# store: Optimized labels conversion on store.Series; Added unsafe labels conversion.