--- a/pkg/store/storepb/custom_test.go
+++ b/pkg/store/storepb/custom_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
+	"github.com/thanos-io/thanos/pkg/store/storepb/prompb"
 	"github.com/thanos-io/thanos/pkg/testutil"
 )
 
@@ -327,4 +328,59 @@ func BenchmarkMergedSeriesSet(b *testing.B) {
 			})
 		}
 	}
+}
+
+var testLsetMap = map[string]string{
+	"a":                           "1",
+	"c":                           "2",
+	"d":                           "dsfsdfsdfsdf123414234",
+	"124134235423534534ffdasdfsf": "1",
+	"":                            "",
+	"b":                           "",
+}
+
+func TestPromLabelsToLabelsUnsafe(t *testing.T) {
+	testutil.Equals(t, PromLabelsToLabels(labels.FromMap(testLsetMap)), PromLabelsToLabelsUnsafe(labels.FromMap(testLsetMap)))
+}
+
+func TestLabelsToPromLabelsUnsafe(t *testing.T) {
+	testutil.Equals(t, labels.FromMap(testLsetMap), LabelsToPromLabels(PromLabelsToLabels(labels.FromMap(testLsetMap))))
+	testutil.Equals(t, labels.FromMap(testLsetMap), LabelsToPromLabelsUnsafe(PromLabelsToLabels(labels.FromMap(testLsetMap))))
+}
+
+func TestPrompbLabelsToLabelsUnsafe(t *testing.T) {
+	var pb []prompb.Label
+	for _, l := range labels.FromMap(testLsetMap) {
+		pb = append(pb, prompb.Label{Name: l.Name, Value: l.Value})
+	}
+	testutil.Equals(t, PromLabelsToLabels(labels.FromMap(testLsetMap)), PrompbLabelsToLabels(pb))
+	testutil.Equals(t, PromLabelsToLabels(labels.FromMap(testLsetMap)), PrompbLabelsToLabelsUnsafe(pb))
+}
+
+func BenchmarkUnsafeVSSafeLabelsConversion(b *testing.B) {
+	const (
+		fmtLbl = "%07daaaaaaaaaabbbbbbbbbbccccccccccdddddddddd"
+		num    = 10000
+	)
+	lbls := make([]labels.Label, 0, num)
+	for i := 0; i < num; i++ {
+		lbls = append(lbls, labels.Label{Name: fmt.Sprintf(fmtLbl, i), Value: fmt.Sprintf(fmtLbl, i)})
+	}
+
+	var converted labels.Labels
+	b.Run("safe", func(b *testing.B) {
+		b.ResetTimer()
+		for i := 0; i < b.N; i++ {
+			converted = LabelsToPromLabels(PromLabelsToLabels(lbls))
+		}
+	})
+	testutil.Equals(b, num, len(converted))
+	b.Run("unsafe", func(b *testing.B) {
+		b.ResetTimer()
+		for i := 0; i < b.N; i++ {
+			converted = LabelsToPromLabelsUnsafe(PromLabelsToLabelsUnsafe(lbls))
+		}
+	})
+	testutil.Equals(b, num, len(converted))
+
 }