## Changes

* method TranslateLables CPU Optimized (streamed sorting).
* All store GW label conversions to []storepb.Label are now alloc-less.


## Benchmark for Unsafe conversion helpers:

```
go test -bench=BenchmarkUnsafeVSSafeLabelsConversion -run=^$ -benchmem -timeout 2h -benchtime 10s ./pkg/store/storepb/...
 goos: linux
 goarch: amd64
 pkg: github.com/thanos-io/thanos/pkg/store/storepb
 BenchmarkUnsafeVSSafeLabelsConversion/safe-12         	   34822	    339076 ns/op	  655368 B/op	       2 allocs/op
 BenchmarkUnsafeVSSafeLabelsConversion/unsafe-12       	1000000000	         2.32 ns/op	       0 B/op	       0 allocs/op
PASS
```

## Series Benchmarks:

Not much difference in memory, meaning the overhead was not much. The latency is much better, most likely because of sorting! (: 

```
benchstat -alpha 100 _dev/bench_outs/man/2.txt _dev/bench_outs/man/new.txt
name                                                     old time/op    new time/op    delta
Series/10e6SeriesWithOneSample/1of10000000-12               20.8s ± 0%     19.7s ± 0%   -5.15%  (p=1.000 n=1+1)
Series/10e6SeriesWithOneSample/10of10000000-12              22.3s ± 0%     21.4s ± 0%   -4.27%  (p=1.000 n=1+1)
Series/10e6SeriesWithOneSample/100of10000000-12             20.6s ± 0%     20.5s ± 0%   -0.55%  (p=1.000 n=1+1)
Series/10e6SeriesWithOneSample/1000of10000000-12            22.7s ± 0%     19.5s ± 0%  -14.18%  (p=1.000 n=1+1)
Series/10e6SeriesWithOneSample/10000of10000000-12           22.5s ± 0%     19.5s ± 0%  -13.23%  (p=1.000 n=1+1)
Series/10e6SeriesWithOneSample/100000of10000000-12          22.1s ± 0%     19.6s ± 0%  -11.14%  (p=1.000 n=1+1)
Series/10e6SeriesWithOneSample/1000000of10000000-12         30.7s ± 0%     25.0s ± 0%  -18.78%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/1of100000000-12            174ms ± 0%     153ms ± 0%  -12.10%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/10of100000000-12           204ms ± 0%     152ms ± 0%  -25.70%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/100of100000000-12          188ms ± 0%     154ms ± 0%  -18.17%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/1000of100000000-12         186ms ± 0%     159ms ± 0%  -14.44%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/10000of100000000-12        175ms ± 0%     153ms ± 0%  -12.56%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/100000of100000000-12       179ms ± 0%     156ms ± 0%  -12.81%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/1000000of100000000-12      248ms ± 0%     185ms ± 0%  -25.46%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/10000000of100000000-12     947ms ± 0%     749ms ± 0%  -20.90%  (p=1.000 n=1+1)

name                                                     old alloc/op   new alloc/op   delta
Series/OneSeriesWith100e6Samples/1of100000000-12           59.3MB ± 0%    59.3MB ± 0%   -0.00%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/10of100000000-12          59.3MB ± 0%    59.3MB ± 0%   -0.00%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/100of100000000-12         59.3MB ± 0%    59.3MB ± 0%   -0.00%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/1000of100000000-12        59.2MB ± 0%    59.2MB ± 0%   -0.00%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/10000of100000000-12       59.3MB ± 0%    59.3MB ± 0%   -0.00%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/100000of100000000-12      60.4MB ± 0%    60.4MB ± 0%   +0.00%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/1000000of100000000-12     9.05MB ± 0%    9.05MB ± 0%   -0.00%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/10000000of100000000-12    97.8MB ± 0%    97.8MB ± 0%   +0.00%  (p=1.000 n=1+1)

name                                                     old allocs/op  new allocs/op  delta
Series/OneSeriesWith100e6Samples/1of100000000-12              253 ± 0%       253 ± 0%     ~     (all equal)
Series/OneSeriesWith100e6Samples/10of100000000-12             253 ± 0%       253 ± 0%     ~     (all equal)
Series/OneSeriesWith100e6Samples/100of100000000-12            252 ± 0%       252 ± 0%     ~     (all equal)
Series/OneSeriesWith100e6Samples/1000of100000000-12           271 ± 0%       271 ± 0%     ~     (all equal)
Series/OneSeriesWith100e6Samples/10000of100000000-12          430 ± 0%       430 ± 0%     ~     (all equal)
Series/OneSeriesWith100e6Samples/100000of100000000-12       1.97k ± 0%     1.97k ± 0%     ~     (all equal)
Series/OneSeriesWith100e6Samples/1000000of100000000-12      17.2k ± 0%     17.2k ± 0%   -0.01%  (p=1.000 n=1+1)
Series/OneSeriesWith100e6Samples/10000000of100000000-12      169k ± 0%      169k ± 0%   +0.01%  (p=1.000 n=1+1)
```

TODO: Do the same on Querier.

Signed-off-by: Bartlomiej Plotka <bwplotka@gmail.com>