@bwplotka 2020-03-08
Some bonus profiles:

```
go test -bench=BenchmarkSeries -run=^$ -benchmem -memprofile .mem10e4series.out -timeout 2h -benchtime 2m ./pkg/store
Starting
goos: linux
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/store
BenchmarkSeries/10e6SeriesWithOneSample/100000of10000000-12         	      45	3019536395 ns/op	1671739740 B/op	20580955 allocs/op
PASS
ok  	github.com/thanos-io/thanos/pkg/store	143.633s
```
![image](https://user-images.githubusercontent.com/6950331/76162234-a0f72800-6133-11ea-83e8-8f52570966a6.png)

```
go test -bench=BenchmarkSeries -run=^$ -benchmem -memprofile .mem10e5series.out -timeout 2h -benchtime 2m ./pkg/store
Starting
goos: linux
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/store
BenchmarkSeries/10e6SeriesWithOneSample/1000000of10000000-12         	      36	3833377516 ns/op	2530649251 B/op	25115272 allocs/op
PASS
ok  	github.com/thanos-io/thanos/pkg/store	146.710s
```
![image](https://user-images.githubusercontent.com/6950331/76162239-aa809000-6133-11ea-9c1e-5e7207b65aee.png)

Diff between querying 10e4 vs 10e5 `go tool pprof -sample_index=alloc_space -http=:8082 -diff_base=.mem10e4series.out .mem10e5series.out`:

![image](https://user-images.githubusercontent.com/6950331/76162259-e7e51d80-6133-11ea-966d-12a84d50ea54.png)