@bwplotka 2020-03-11
Definitely extra test will be nice here :heart: 

Will do