@bwplotka 2020-03-16
Merged, but will add tests in next PR @GiedriusS