@bwplotka 2020-03-27
@GiedriusS did not forgot about this. Adding test like this:

```
func TestLabelsToPrompbCompatibility(t *testing.T) {
	l := labels.Label{Name: "name", Value: "value"}

	p := (*prompb.Label)(unsafe.Pointer(&l))

	testutil.Equals(t, unsafe.Sizeof(&l), unsafe.Sizeof(p))
	testutil.Equals(t, prompb.Label{Name: "name", Value: "value"}, *p)

	// Just to show that memory is reused.
	p.Value = "changed"
	testutil.Equals(t, prompb.Label{Name: l.Name, Value: l.Value}, *p)
}
```