@GiedriusS 2020-04-02
> @GiedriusS did not forgot about this. Adding test like this:
> 
> ```
> func TestLabelsToPrompbCompatibility(t *testing.T) {
> 	l := labels.Label{Name: "name", Value: "value"}
> 
> 	p := (*prompb.Label)(unsafe.Pointer(&l))
> 
> 	testutil.Equals(t, unsafe.Sizeof(&l), unsafe.Sizeof(p))
> 	testutil.Equals(t, prompb.Label{Name: "name", Value: "value"}, *p)
> 
> 	// Just to show that memory is reused.
> 	p.Value = "changed"
> 	testutil.Equals(t, prompb.Label{Name: l.Name, Value: l.Value}, *p)
> }
> ```

@bwplotka 

![beanwaiting](https://user-images.githubusercontent.com/2377233/78269748-aa956500-7512-11ea-88eb-040687e5957c.gif)