@GiedriusS 2020-05-10
> > @GiedriusS did not forgot about this. Adding test like this:
> > ```
> > func TestLabelsToPrompbCompatibility(t *testing.T) {
> > 	l := labels.Label{Name: "name", Value: "value"}
> > 
> > 	p := (*prompb.Label)(unsafe.Pointer(&l))
> > 
> > 	testutil.Equals(t, unsafe.Sizeof(&l), unsafe.Sizeof(p))
> > 	testutil.Equals(t, prompb.Label{Name: "name", Value: "value"}, *p)
> > 
> > 	// Just to show that memory is reused.
> > 	p.Value = "changed"
> > 	testutil.Equals(t, prompb.Label{Name: l.Name, Value: l.Value}, *p)
> > }
> > ```

@bwplotka 

![waiting](https://user-images.githubusercontent.com/2377233/81497767-44e28880-92c9-11ea-9109-a8177aaed07c.gif)